streamlit/v2/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
altair==5.3.0
aniso8601==9.0.1
annotated-types==0.7.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.3
attrs==23.2.0
awswrangler==3.9.1
blinker==1.8.2
blis==0.7.11
boto3==1.34.149
botocore==1.34.149
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
cloudpickle==2.2.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
databricks-sdk==0.29.0
dataclasses-json==0.6.7
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
Deprecated==1.2.14
dill==0.3.8
docker==7.1.0
entrypoints==0.4
Events==0.5
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1720869315914/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.33.0
google-pasta==0.2.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
gunicorn==22.0.0
huggingface-hub==0.24.5
idna==3.7
importlib-metadata==6.11.0
importlib_resources==6.4.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1701831663892/work
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-ng==1.6.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1716472197302/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1710257447442/work
kiwisolver==1.4.5
langchain==0.2.11
langchain-aws==0.1.12
langchain-community==0.2.10
langchain-core==0.2.24
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.2
langcodes==3.4.0
langsmith==0.1.93
language_data==1.2.0
Mako==1.3.5
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1.post1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdurl==0.1.2
mlflow==2.15.1
mlflow-skinny==2.15.1
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
opensearch-py==2.6.0
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
orjson==3.10.6
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==2.2.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pathos==0.3.2
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.4.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
ply==3.11
pox==0.3.4
ppft==1.7.6.8
preshed==3.0.9
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
protobuf==4.25.4
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1719274564771/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pyparsing==3.1.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytz==2024.1
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1705605076900/work
querystring-parser==1.2.4
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.4
sagemaker==2.226.1
schema==0.7.7
scikit-learn==1.5.1
scipy==1.13.1
sentence-transformers==3.0.1
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open==7.0.4
smdebug-rulesconfig==1.0.1
smmap==5.0.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-llm==0.7.2
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
sqlparse==0.5.1
srsly==2.4.8
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
streamlit==1.37.1
sympy==1.13.1
tblib==3.0.0
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.4.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1717722887379/work
tqdm==4.66.4
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
transformers==4.44.0
triton==3.0.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.1
urllib3==1.26.19
wasabi==1.1.3
watchdog==4.0.1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
weasel==0.4.1
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1718013267051/work
